feat(auth): Remove auto admin for first user
The previous logic for automatically promoting the first registered user to admin and approving their status has been removed. This caused permission errors during registration as it required read access to the users collection before authentication.

The first user will now need to be manually set as admin and approved in the Firebase console.