feat: Add "Remember Me" functionality to login
Implements a "Remember Me" checkbox on the login page.
When checked, the user's email is stored in local storage, persisting across sessions.
On subsequent visits, if "Remember Me" was previously checked, the email field will be pre-populated.
The email is cleared from local storage if the user unchecks the box or logs out.