feat: Add open in new tab for ID card
Introduces a new feature to open the virtual student ID card in a new browser tab. This provides users with an alternative way to view and interact with their ID card without the need for downloading a PDF.

Also includes minor UI adjustments and optimizations for the StudentIdCard component, such as removing unnecessary `crossOrigin` attribute.